llm-service-v2/requirements.txt
fastapi==0.115.11
uvicorn==0.34.0
openai==1.75.0
pydantic==2.10.6
langgraph==0.4.1
langchain==0.3.24
langchain-openai==0.3.15
langchain-core==0.3.56
python-chess==1.999
tiktoken==0.9.0